Moved atram to package
